fix(quotes): OQ-02 review — per-req send-activity id + fix service→router layering
Silent partial-failure finding: send_quote_email logged each contributing
requisition's send-activity with the SAME external_id=graph_message_id, and
log_email_activity dedupes by external_id — so in production (non-null Graph id)
only the PRIMARY req got its send ActivityLog and every secondary req was
silently dropped. Fix: keep the PRIMARY on the raw graph id (the sent-folder
reconcile in email_jobs matches the row by that exact id and would otherwise
create a duplicate) and suffix each SECONDARY req's id (":req{rid}") so every
contributing req records its own send. The existing test masked this by using
testing=True (graph_message_id stays None → dedup off); it now drives the real
send path with a patched Graph so a shared id would actually drop the second req.

Layering fix: moved _customer_name_for_site from routers/quote_builder into
services/quote_requisitions (the service was importing UP into the router); the
router now imports it from the service (router → service, per CLAUDE.md).
